+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/AuthorizationHandler.java
@@ -33,7 +33,7 @@ import co.cask.cdap.proto.security.Principal;
 import co.cask.cdap.proto.security.Privilege;
 import co.cask.cdap.proto.security.RevokeRequest;
 import co.cask.cdap.proto.security.Role;
-import co.cask.cdap.security.authorization.AuthorizerInstantiatorService;
+import co.cask.cdap.security.authorization.AuthorizerSupplier;
 import co.cask.cdap.security.spi.authentication.SecurityRequestContext;
 import co.cask.cdap.security.spi.authorization.Authorizer;
 import co.cask.http.HttpResponder;
@@ -67,7 +67,7 @@ import javax.ws.rs.PathParam;
 public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
 
   private static final Logger AUDIT_LOG = LoggerFactory.getLogger("authorization-access");
-  private final AuthorizerInstantiatorService authorizerInstantiatorService;
+  private final AuthorizerSupplier authorizerSupplier;
   private static final Gson GSON = new GsonBuilder()
     .registerTypeAdapter(EntityId.class, new EntityIdTypeAdapter())
     .create();
@@ -77,9 +77,9 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
   private final EntityExistenceVerifier entityExistenceVerifier;
 
   @Inject
-  AuthorizationHandler(AuthorizerInstantiatorService authorizerInstantiatorService, CConfiguration cConf,
+  AuthorizationHandler(AuthorizerSupplier authorizerSupplier, CConfiguration cConf,
                        EntityExistenceVerifier entityExistenceVerifier) {
-    this.authorizerInstantiatorService = authorizerInstantiatorService;
+    this.authorizerSupplier = authorizerSupplier;
     this.authenticationEnabled = cConf.getBoolean(Constants.Security.ENABLED);
     this.authorizationEnabled = cConf.getBoolean(Constants.Security.Authorization.ENABLED);
     this.entityExistenceVerifier = entityExistenceVerifier;
@@ -95,9 +95,9 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
 
     Set<Action> actions = request.getActions() == null ? EnumSet.allOf(Action.class) : request.getActions();
     // enforce that the user granting access has admin privileges on the entity
-    authorizerInstantiatorService.get().enforce(request.getEntity(), SecurityRequestContext.toPrincipal(),
+    authorizerSupplier.get().enforce(request.getEntity(), SecurityRequestContext.toPrincipal(),
                                                 Action.ADMIN);
-    authorizerInstantiatorService.get().grant(request.getEntity(), request.getPrincipal(), actions);
+    authorizerSupplier.get().grant(request.getEntity(), request.getPrincipal(), actions);
 
     httpResponder.sendStatus(HttpResponseStatus.OK);
     createLogEntry(httpRequest, request, HttpResponseStatus.OK);
@@ -112,13 +112,13 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
     verifyAuthRequest(request);
 
     // enforce that the user revoking access has admin privileges on the entity
-    authorizerInstantiatorService.get().enforce(request.getEntity(), SecurityRequestContext.toPrincipal(),
+    authorizerSupplier.get().enforce(request.getEntity(), SecurityRequestContext.toPrincipal(),
                                                 Action.ADMIN);
     if (request.getPrincipal() == null && request.getActions() == null) {
-      authorizerInstantiatorService.get().revoke(request.getEntity());
+      authorizerSupplier.get().revoke(request.getEntity());
     } else {
       Set<Action> actions = request.getActions() == null ? EnumSet.allOf(Action.class) : request.getActions();
-      authorizerInstantiatorService.get().revoke(request.getEntity(), request.getPrincipal(), actions);
+      authorizerSupplier.get().revoke(request.getEntity(), request.getPrincipal(), actions);
     }
 
     httpResponder.sendStatus(HttpResponseStatus.OK);
@@ -132,7 +132,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
                              @PathParam("principal-name") String principalName) throws Exception {
     ensureSecurityEnabled();
     Principal principal = new Principal(principalName, Principal.PrincipalType.valueOf(principalType.toUpperCase()));
-    httpResponder.sendJson(HttpResponseStatus.OK, authorizerInstantiatorService.get().listPrivileges(principal),
+    httpResponder.sendJson(HttpResponseStatus.OK, authorizerSupplier.get().listPrivileges(principal),
                            PRIVILEGE_SET_TYPE, GSON);
     createLogEntry(httpRequest, null, HttpResponseStatus.OK);
   }
@@ -147,7 +147,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
   public void createRole(HttpRequest httpRequest, HttpResponder httpResponder,
                          @PathParam("role-name") String roleName) throws Exception {
     ensureSecurityEnabled();
-    authorizerInstantiatorService.get().createRole(new Role(roleName));
+    authorizerSupplier.get().createRole(new Role(roleName));
     httpResponder.sendStatus(HttpResponseStatus.OK);
     createLogEntry(httpRequest, null, HttpResponseStatus.OK);
   }
@@ -157,7 +157,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
   public void dropRole(HttpRequest httpRequest, HttpResponder httpResponder,
                        @PathParam("role-name") String roleName) throws Exception {
     ensureSecurityEnabled();
-    authorizerInstantiatorService.get().dropRole(new Role(roleName));
+    authorizerSupplier.get().dropRole(new Role(roleName));
     httpResponder.sendStatus(HttpResponseStatus.OK);
     createLogEntry(httpRequest, null, HttpResponseStatus.OK);
   }
@@ -166,7 +166,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
   @GET
   public void listAllRoles(HttpRequest httpRequest, HttpResponder httpResponder) throws Exception {
     ensureSecurityEnabled();
-    httpResponder.sendJson(HttpResponseStatus.OK, authorizerInstantiatorService.get().listAllRoles());
+    httpResponder.sendJson(HttpResponseStatus.OK, authorizerSupplier.get().listAllRoles());
     createLogEntry(httpRequest, null, HttpResponseStatus.OK);
   }
 
@@ -177,7 +177,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
                         @PathParam("principal-name") String principalName) throws Exception {
     ensureSecurityEnabled();
     Principal principal = new Principal(principalName, Principal.PrincipalType.valueOf(principalType.toUpperCase()));
-    httpResponder.sendJson(HttpResponseStatus.OK, authorizerInstantiatorService.get().listRoles(principal));
+    httpResponder.sendJson(HttpResponseStatus.OK, authorizerSupplier.get().listRoles(principal));
     createLogEntry(httpRequest, null, HttpResponseStatus.OK);
   }
 
@@ -188,7 +188,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
                                  @PathParam("principal-name") String principalName,
                                  @PathParam("role-name") String roleName) throws Exception {
     ensureSecurityEnabled();
-    authorizerInstantiatorService.get().addRoleToPrincipal(new Role(roleName),
+    authorizerSupplier.get().addRoleToPrincipal(new Role(roleName),
                                   new Principal(principalName,
                                                 Principal.PrincipalType.valueOf(principalType.toUpperCase())));
     httpResponder.sendStatus(HttpResponseStatus.OK);
@@ -202,7 +202,7 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
                                       @PathParam("principal-name") String principalName,
                                       @PathParam("role-name") String roleName) throws Exception {
     ensureSecurityEnabled();
-    authorizerInstantiatorService.get().removeRoleFromPrincipal(new Role(roleName),
+    authorizerSupplier.get().removeRoleFromPrincipal(new Role(roleName),
                                        new Principal(principalName,
                                                      Principal.PrincipalType.valueOf(principalType.toUpperCase())));
     httpResponder.sendStatus(HttpResponseStatus.OK);
