+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/namespace/DefaultNamespaceAdmin.java
@@ -45,7 +45,7 @@ import co.cask.cdap.proto.id.InstanceId;
 import co.cask.cdap.proto.id.NamespaceId;
 import co.cask.cdap.proto.security.Action;
 import co.cask.cdap.proto.security.Principal;
-import co.cask.cdap.security.authorization.AuthorizerInstantiatorService;
+import co.cask.cdap.security.authorization.AuthorizerSupplier;
 import co.cask.cdap.security.spi.authentication.SecurityRequestContext;
 import co.cask.cdap.store.NamespaceStore;
 import com.google.common.base.Preconditions;
@@ -80,7 +80,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
   private final Scheduler scheduler;
   private final ApplicationLifecycleService applicationLifecycleService;
   private final ArtifactRepository artifactRepository;
-  private final AuthorizerInstantiatorService authorizerInstantiatorService;
+  private final AuthorizerSupplier authorizerSupplier;
   private final InstanceId instanceId;
   private final Pattern namespacePattern = Pattern.compile("[a-zA-Z0-9_]+");
 
@@ -91,7 +91,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
                         MetricStore metricStore, Scheduler scheduler,
                         ApplicationLifecycleService applicationLifecycleService,
                         ArtifactRepository artifactRepository,
-                        AuthorizerInstantiatorService authorizerInstantiatorService,
+                        AuthorizerSupplier authorizerSupplier,
                         CConfiguration cConf) {
     this.queueAdmin = queueAdmin;
     this.streamAdmin = streamAdmin;
@@ -105,7 +105,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     this.metricStore = metricStore;
     this.applicationLifecycleService = applicationLifecycleService;
     this.artifactRepository = artifactRepository;
-    this.authorizerInstantiatorService = authorizerInstantiatorService;
+    this.authorizerSupplier = authorizerSupplier;
     this.instanceId = createInstanceId(cConf);
   }
 
@@ -171,7 +171,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     // Skip authorization enforcement for the system user and the default namespace, so the DefaultNamespaceEnsurer
     // thread can successfully create the default namespace
     if (!(Principal.SYSTEM.equals(principal) && NamespaceId.DEFAULT.equals(namespace))) {
-      authorizerInstantiatorService.get().enforce(instanceId, principal, Action.ADMIN);
+      authorizerSupplier.get().enforce(instanceId, principal, Action.ADMIN);
     }
 
     try {
@@ -183,7 +183,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     nsStore.create(metadata);
     // Skip authorization grants for the system user
     if (!(Principal.SYSTEM.equals(principal) && NamespaceId.DEFAULT.equals(namespace))) {
-      authorizerInstantiatorService.get().grant(namespace, principal, ImmutableSet.of(Action.ALL));
+      authorizerSupplier.get().grant(namespace, principal, ImmutableSet.of(Action.ALL));
     }
   }
 
@@ -210,8 +210,8 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     }
 
     // Namespace can be deleted. Revoke all privileges first
-    authorizerInstantiatorService.get().enforce(namespace, SecurityRequestContext.toPrincipal(), Action.ADMIN);
-    authorizerInstantiatorService.get().revoke(namespace);
+    authorizerSupplier.get().enforce(namespace, SecurityRequestContext.toPrincipal(), Action.ADMIN);
+    authorizerSupplier.get().revoke(namespace);
 
     LOG.info("Deleting namespace '{}'.", namespaceId);
     try {
@@ -277,7 +277,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     }
 
     // Namespace data can be deleted. Revoke all privileges first
-    authorizerInstantiatorService.get().enforce(namespaceId.toEntityId(), SecurityRequestContext.toPrincipal(),
+    authorizerSupplier.get().enforce(namespaceId.toEntityId(), SecurityRequestContext.toPrincipal(),
                                                 Action.ADMIN);
     try {
       dsFramework.deleteAllInstances(namespaceId);
@@ -293,7 +293,7 @@ public final class DefaultNamespaceAdmin implements NamespaceAdmin {
     if (!exists(namespaceId)) {
       throw new NamespaceNotFoundException(namespaceId);
     }
-    authorizerInstantiatorService.get().enforce(namespaceId.toEntityId(), SecurityRequestContext.toPrincipal(),
+    authorizerSupplier.get().enforce(namespaceId.toEntityId(), SecurityRequestContext.toPrincipal(),
                                                 Action.ADMIN);
     NamespaceMeta metadata = nsStore.get(namespaceId);
     NamespaceMeta.Builder builder = new NamespaceMeta.Builder(metadata);
