+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/services/ApplicationLifecycleService.java
@@ -67,7 +67,7 @@ import co.cask.cdap.proto.id.ApplicationId;
 import co.cask.cdap.proto.id.Ids;
 import co.cask.cdap.proto.id.ProgramId;
 import co.cask.cdap.proto.security.Action;
-import co.cask.cdap.security.authorization.AuthorizerInstantiatorService;
+import co.cask.cdap.security.authorization.AuthorizerSupplier;
 import co.cask.cdap.security.spi.authentication.SecurityRequestContext;
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
@@ -123,7 +123,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
   private final ArtifactRepository artifactRepository;
   private final ManagerFactory<AppDeploymentInfo, ApplicationWithPrograms> managerFactory;
   private final MetadataStore metadataStore;
-  private final AuthorizerInstantiatorService authorizerInstantiatorService;
+  private final AuthorizerSupplier authorizerSupplier;
 
   @Inject
   ApplicationLifecycleService(ProgramRuntimeService runtimeService, Store store, CConfiguration configuration,
@@ -134,7 +134,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
                               ArtifactRepository artifactRepository,
                               ManagerFactory<AppDeploymentInfo, ApplicationWithPrograms> managerFactory,
                               MetadataStore metadataStore,
-                              AuthorizerInstantiatorService authorizerInstantiatorService) {
+                              AuthorizerSupplier authorizerSupplier) {
     this.runtimeService = runtimeService;
     this.store = store;
     this.configuration = configuration;
@@ -148,7 +148,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
     this.artifactRepository = artifactRepository;
     this.managerFactory = managerFactory;
     this.metadataStore = metadataStore;
-    this.authorizerInstantiatorService = authorizerInstantiatorService;
+    this.authorizerSupplier = authorizerSupplier;
   }
 
   @Override
@@ -240,7 +240,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
     }
     // App exists. Check if the current user has admin privileges on it before updating. The user's write privileges on
     // the namespace will get enforced in the deployApp method.
-    authorizerInstantiatorService.get().enforce(appId.toEntityId(), SecurityRequestContext.toPrincipal(), Action.ADMIN);
+    authorizerSupplier.get().enforce(appId.toEntityId(), SecurityRequestContext.toPrincipal(), Action.ADMIN);
     ArtifactId currentArtifact = currentSpec.getArtifactId();
 
     // if no artifact is given, use the current one.
@@ -534,7 +534,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
                                             ProgramTerminator programTerminator,
                                             ArtifactDetail artifactDetail) throws Exception {
     // Enforce that the current principal has write access to the namespace the app is being deployed to
-    authorizerInstantiatorService.get().enforce(namespace.toEntityId(), SecurityRequestContext.toPrincipal(),
+    authorizerSupplier.get().enforce(namespace.toEntityId(), SecurityRequestContext.toPrincipal(),
                                                 Action.WRITE);
     Id.Artifact artifactId = Id.Artifact.from(namespace, artifactDetail.getDescriptor().getArtifactId());
     Set<ApplicationClass> appClasses = artifactDetail.getMeta().getClasses().getApps();
@@ -551,7 +551,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
     // TODO: (CDAP-3258) Manager needs MUCH better error handling.
     ApplicationWithPrograms applicationWithPrograms = manager.deploy(namespace, appName, deploymentInfo).get();
     // Deployment successful. Grant all privileges on this app to the current principal.
-    authorizerInstantiatorService.get().grant(namespace.toEntityId().app(applicationWithPrograms.getId().getId()),
+    authorizerSupplier.get().grant(namespace.toEntityId().app(applicationWithPrograms.getId().getId()),
                                               SecurityRequestContext.toPrincipal(), ImmutableSet.of(Action.ALL));
     return applicationWithPrograms;
   }
@@ -567,7 +567,7 @@ public class ApplicationLifecycleService extends AbstractIdleService {
    */
   private void deleteApp(Id.Application appId, ApplicationSpecification spec) throws Exception {
     // enfore ADMIN privileges on the app
-    authorizerInstantiatorService.get().enforce(appId.toEntityId(), SecurityRequestContext.toPrincipal(), Action.ADMIN);
+    authorizerSupplier.get().enforce(appId.toEntityId(), SecurityRequestContext.toPrincipal(), Action.ADMIN);
     // first remove all privileges on the app
     revokePrivileges(appId.toEntityId(), spec);
 
@@ -621,9 +621,9 @@ public class ApplicationLifecycleService extends AbstractIdleService {
   // TODO: CDAP-5427 - This should be a single operation
   private void revokePrivileges(ApplicationId appId, ApplicationSpecification appSpec) throws Exception {
     for (ProgramId programId : getAllPrograms(appId, appSpec)) {
-      authorizerInstantiatorService.get().revoke(programId);
+      authorizerSupplier.get().revoke(programId);
     }
-    authorizerInstantiatorService.get().revoke(appId);
+    authorizerSupplier.get().revoke(appId);
   }
 
   /**
